test/inspector/console/console-follow-spec-expected.txt

Running test: consoleExistsOnGlobal
true

Running test: consoleHasRightPropertyDescriptor
{
    configurable : true
    enumerable : false
    value : <value>
    writable : true
}

Running test: ConsoleNotExistsOnGlobal
false

Running test: prototypeChainMustBeCorrect
true
